feat(#1225): move swap state types
