+++ src/state/swap/reducer.ts
@@ -1,29 +1,7 @@
 import { createReducer } from '@reduxjs/toolkit'
 
-import {
-  Field,
-  replaceSwapState,
-  selectCurrency,
-  setLoading,
-  setRecipient,
-  switchCurrencies,
-  typeInput,
-} from './actions'
-
-export interface SwapState {
-  readonly independentField: Field
-  readonly typedValue: string
-  readonly [Field.INPUT]: {
-    readonly currencyId: string | undefined
-  }
-  readonly [Field.OUTPUT]: {
-    readonly currencyId: string | undefined
-  }
-  // the typed recipient address or ENS name, or null if swap should go to sender
-  readonly recipient: string | null
-  readonly protocolFeeTo: string | undefined
-  readonly loading: boolean
-}
+import { replaceSwapState, selectCurrency, setLoading, setRecipient, switchCurrencies, typeInput } from './actions'
+import { Field, SwapState } from './types'
 
 export const initialState: SwapState = {
   independentField: Field.INPUT,
