+++ src/state/swap/types.ts
@@ -1,0 +1,19 @@
+export enum Field {
+  INPUT = 'INPUT',
+  OUTPUT = 'OUTPUT',
+}
+
+export interface SwapState {
+  readonly independentField: Field
+  readonly typedValue: string
+  readonly [Field.INPUT]: {
+    readonly currencyId: string | undefined
+  }
+  readonly [Field.OUTPUT]: {
+    readonly currencyId: string | undefined
+  }
+  // the typed recipient address or ENS name, or null if swap should go to sender
+  readonly recipient: string | null
+  readonly protocolFeeTo: string | undefined
+  readonly loading: boolean
+}
